refactor(auth): move session/token hooks from lib/users to lib/auth
